MGMT-2977 Cluster specific minimal iso (#932)
* MGMT-297 Add discovery image type

This adds a type enum to both the cluster image info and the
image creation parameters

* MGMT-2977 Move tempfile handling into create()

* MGMT-2977 Move logic for creating ignition archive to isoeditor

This will make it easier to share between the existing logic
which creates the archive then uploads it to s3 and the new logic
which will write the archive to the fs before bundling it into the
minimal iso directly

* MGMT-2977 Add the full ignition to the iso before re-creating it

This allows us to create the entire iso locally and skip the process
of uploading the ignition to S3

* MGMT-2977 Create a cluster-specific minimal iso when requested

* MGMT-2977 Create a factory interface for isoeditor

This will be a good place to add "ratelimiting" for concurrent iso
editing if that's something we need.

Additionally it makes things easier to test by giving us somewhere
to inject mock objects at the inventory API level.

Authored-by: Nick Carboni <ncarboni@redhat.com>

* MGMT-2977 extract generate cluster minimal iso logic

Extracted generate cluster minimal ISO logic to 'generateClusterMinimalISO'
function in order to avoid a cyclomatic complexity error.

* MGMT-2977 negative unit-tests for generate minimal iso

* MGMT-2977 unit-test for regenerating iso template

Added "Regenerates the iso for a new type" unit-test.

* MGMT-2977 added minimal-iso ImageType to subsystem tests

Co-authored-by: Daniel Erez <derez@redhat.com>