MGMT-4942 Download rootfs from mirror.openshift.com (take 2) (#1697)
* MGMT-4942 Add rootfs url to openshift version model

By default this uses the rootfs published with the rhcos images
This will need to be overridden to use an internally hosted url for
disconnected environments

* MGMT-4942 Remove openshift version from WithEditor

The version is only needed when creating the minimal iso template
so we can pass it as a parameter to that function instead of requiring
it for every editor instance.

* MGMT-4942 Use rootfs url from versions rather than our service

This passes the url through to the isoeditor when creating the minimal
iso template and embeds that rather than a url pointing back to the
service boot-files endpoint.

This also increases the minimal iso template version number to force a
regeneration.