Add setup script to share logic with prow (#6)
Both the prow test container and our build container will need to
install the same test dependencies and we want to keep those as similar
as possible.

After this is merged, the prow test-bin container will call the new
`build-openshift-ci-test-bin` make target to ensure our dependencies are
installed

This also adds the `genisoimage` command in preparation of the tests
from the iso editing code.