Remove system area metadata embedding
In assisted-service, the base images were stored in a remote system,
often some S3 compatible storage. This meant that we needed a
convienient way to read the location of the ignition.

Here, we will have all the base images locally. This means that we can
read the file location directly instead of having to embed that location
in some special area.

The extra embedding is just leaving us open to bugs because the logic is
rather complex. With this solution we're doing the same thing we would
have previously to find the information we need to embed, just every
time we need to customize an iso.

https://issues.redhat.com/browse/MGMT-7256