MGMT-7256 Add ISO stream customization (#8)
* Remove system area metadata embedding

In assisted-service, the base images were stored in a remote system,
often some S3 compatible storage. This meant that we needed a
convienient way to read the location of the ignition.

Here, we will have all the base images locally. This means that we can
read the file location directly instead of having to embed that location
in some special area.

The extra embedding is just leaving us open to bugs because the logic is
rather complex. With this solution we're doing the same thing we would
have previously to find the information we need to embed, just every
time we need to customize an iso.

https://issues.redhat.com/browse/MGMT-7256

* Combine isoutil file info functions

We don't need this info separately so we should only open the file once
to get both the size and offset

https://issues.redhat.com/browse/MGMT-7256

* Add the streaming iso reader

This commits adds the overlay reader implementation from
https://github.com/zaneb/iso-stream, but provides the offset information
directly from the iso file rather than reading it from the embed area.

https://issues.redhat.com/browse/MGMT-7256

* Remove unnecessary -self_package option from rhcos mockgen

* Add tests for parsing out the image ID

Previously this wasn't capturing the image id, but was matching and
returning the whole string. Also fix the function name.

https://issues.redhat.com/browse/MGMT-7256

* Embed ignition from assisted service into iso stream

This commit queries assisted service for the ignition using the
provided id as the cluster id. That ignition is embedded into the
base image using the stream reader.

https://issues.redhat.com/browse/MGMT-7256

* Deal with assisted-service API correctly

This commit fixes the assisted service endpoint and properly creates the
compresed archive to embed into the iso.

https://issues.redhat.com/browse/MGMT-7256