feat: support self-signed certificates on assisted-service (#19)
* chore: make README use valid json

* feat: add client to imagehandler

The assisted-service may likely be using self-signed certificates. This
change makes it possible to start the imagehandler with a cabundle to
make trust between the image-service and assisted-service possible.

Additionally, we break up the handling of `err != nil` to ensure we
don't panic because response != 2xx but err was nil.

* test: update tests with default client

* fix: use a newer tls version

* NO-ISSUE: Address review comments