Add support for multi-arch base images (#18)
* Refactor image store tests

Some of these didn't need to be in the larger contexts

* Use arch-specific URLs in default versions

If we have multiple architectures in here eventually it will be better
for them to follow the same URL pattern rather than some relying on
being the default path

* Add multi arch support to the API

This commit adds a new parameter to the API that indicates the cpu
architecture of the base image to be used.

The arch and version parameters are used together to determine which
base image will be selected and that combination must be present in the
versions list.

https://issues.redhat.com/browse/MGMT-7453

* Add RHCOS_VERSIONS to `make run`

* Don't log the whole versions map in info messages

That level of detail might be good in the error message, but it's
too much for info