MGMT-7734: Integration tests for the image-service  (#28)
* Take version map as parameter to NewImageStore

Moving the handling of the `RHCOS_VERSIONS` json string to main makes it
possible for us to instantiate an ImageStore without relying on the
RHCOS_VERSIONS environment variable.

This also allows us to test the image-service without needing to muck
with the host environment.

* Integration tests for the image-service

Verify that we can grab a few versions (and architectures) of RHCOS from
the openshift mirror and:
1. Create full-iso, minimal-iso template
2. Stream an RHCOS image with ignition + initrd customizations embedded
   properly

Since this test could take a long time (and longer in the future as we
support more versions and architectures), we'll skip this when run using
`go test -short`. So this change also includes a new make target
`test-integration` to run this test.

* Don't use default prometheus.Registry

Instead of relying on Prometheus' global default registry, we'll
instatiate one and provide it to the handlers as necessary.

- Take registry as argument on the NewImageHandler
- Update integration tests

* move (Before|After)Suite to suite_test

* fix loop and closure issue

* refactor integration_test

* remove config from imagestore

* refactor to remove duplication in integration test