Bug 2010819: Use the request context when making requests to assisted-service (#29)
* Use the golang.org semaphore package

Don't implement our own semaphore. It will be more reliable to use the
one from golang.org.

* Use the request context when making requests to assisted-service

After this change requests to assisted-service will be
canceled if/when the request to the image service is canceled.

This will ensure that requests that the client doesn't care about won't
run nmstatectl processes on assisted-service.

https://bugzilla.redhat.com/show_bug.cgi?id=2010819