Make Ignition archiving code public (#35)
All users of NewRHCOSStreamReader will need to convert the image data to
a gzip-compressed CPIO archive, so move this code out of the internal
package. To ensure future compatibility (e.g. if more than one file had
to be inserted in the archive), implement this code in a new
IgnitionContent type.