Use the version key for file naming
This is required for when we update the configuration to point to a new
z-stream image. Previously only the X.Y version was used for the image
path, but now that the data directory could be persisted across runs a
more exact strategy is required for naming the template image files.

https://issues.redhat.com/browse/MGMT-8216