MGMT-8216: Handle persistent storage for the data directory (#39)
* Clean old files from the data directory on startup

This commit determines the list of file names that we expect in the data
directory and removes anything not on that list before downloading and
creating the new template images.

This commit also intentionally does not add the minimal iso file names
to the list of allowed files. This is to force them to be regenerated on
each deploy so that we can fix bugs that affect the minimal iso creation
process without also forcing users to remove those images to get them to
regenerate.

https://issues.redhat.com/browse/MGMT-8216

* Update the default versions and add the 'version' key

https://issues.redhat.com/browse/MGMT-8216

* Use the version key for file naming

This is required for when we update the configuration to point to a new
z-stream image. Previously only the X.Y version was used for the image
path, but now that the data directory could be persisted across runs a
more exact strategy is required for naming the template image files.

https://issues.redhat.com/browse/MGMT-8216