Add the ability to customise PXE files (#42)
* overlay: Refactor NewOverlayReader

* overlay: Add NewAppendReader function

Provide a convenience function for creating a stream that just appends
one stream after another.

* isoeditor: Add NewInitRamFSStreamReader()

Add a function to create an initramfs stream from an initramfs file and
IgnitionContent.

The format of the initramfs is a series of concatenated gzipped cpio
archives, so the customisation consists simply of appending the same
Ignition archive used for ISO customisation to the input initramfs file:

https://www.kernel.org/doc/html/latest/driver-api/early-userspace/buffer-format.html