Add an initrd handler to serve custom initrd
This handler is also rooted at `/images/:imageID` but then uses the
`pxe-initrd` sub-path. This means that we need a single handler for
`/images` which routes to the inner handler based on if the url is to an
image or the initrd.

The initrd handler serves a the initrd from the requested iso with the
ignition archive appended.

https://issues.redhat.com/browse/MGMT-9333