Fix overlay reader for base readers that return EOF early (#63)
It's valid for a reader to return EOF at the same time as the last
bytes. If we have such a reader as the base for an overlay we return EOF
when the base reader returns EOF even if we haven't printed the overlay
yet.

This is specifically a problem for the append reader when we're always
getting to the end of the base before starting to read the overlay.

https://issues.redhat.com/browse/MGMT-9333