MGMT-9332: Add endpoint for rootfs and kernel (#62)
* MGMT-9332: Add endpoint for rootfs and kernel.

As part of the larger epic to serve images with iPXE
https://issues.redhat.com/browse/MGMT-8580

This PR will add a new endpoint to the image service
which will serve the rootfs or the kernel directly
from an ISO.

* Fix boot artifacts tests

Remove a circular import, and don't define TestHandlers multiple times

* Improve boot artifact error handling

Ensure we have the requsted version before trying to serve a file from
it and fail if we can't stat the base iso file.

* Fix temp dir leak in boot artifacts test

Previously this was creating a new directory in the os temp dir for
every test in this context.

* Add metrics and CORS for boot-artifacts endpoint

Move the metrics setup out of the image handler because it needs to be
shared between all endpoints.

* Allow HEAD requests for boot-artifacts

* Return context with error responses in boot-artifacts handler

Co-authored-by: CrystalChun <cchun@redhat.com>