MGMT-9333 Add ipxe-initrd endpoint for customized initrd (#64)
* Move assisted service querying out of the image handler

This functionality will need to be reused for fetching the ignition when
serving the initrd so it can't be buried in the image handler.

https://issues.redhat.com/browse/MGMT-9333

* Create a separate iso handler from the image handler

The image handler will be the entry point for paths rooted at `/images/`
but we will need serve both isos and initrds from there so the existing
images handler will act as a router to those two sub-paths.

* Add an initrd handler to serve custom initrd

This handler is also rooted at `/images/:imageID` but then uses the
`pxe-initrd` sub-path. This means that we need a single handler for
`/images` which routes to the inner handler based on if the url is to an
image or the initrd.

The initrd handler serves a the initrd from the requested iso with the
ignition archive appended.

https://issues.redhat.com/browse/MGMT-9333

* Add pxe-initrd endpoint documentation in README

* Add a http error handling helper function

This makes a ServeHTTP implementation much cleaner when we have to both
log and return a message in the http response during error handling.