Return a 503 from images and boot-artifacts before downloads finish
This commit extends the readiness handler to provide a middleware that
will return 503s until we mark the handler ready. It will then call the
underlying handler.

This new middleware is applied to both the images and boot-artifacts
endpoints so it's more clear to a user that the failure it temporary
rather than some issue with the deployment or configuration.

https://issues.redhat.com/browse/MGMT-9395