Set Last-Modified time for iso and initrd downloads (#72)
This time is based on the last modified time sent to the service when
querying for the ignition from assisted-service.

This will be used by the ironic cache to determine when it needs to
re-download an image. Before this change the image would be
re-downloaded every time a host booted, but now the cache will start
acting like a cache.

https://issues.redhat.com/browse/MGMT-7274