MGMT-10751: use boot-artifacts API as rootfs URL
When creating a minimal ISO template the rootfsURL is injected
to grub config files. For each OS image metadata (OS_IMAGES)
the rootfs_url is available as an external URL to the img file.
This URL is currently used on imagestore -> Populate.

This change is removing the requirement to specify an external
rootfs URL for each image, instead, we're now using the
/boot-artifacts/rootfs API. I.e. a URL as follows is generated
when creating the minimal ISO and injected to the grub configs:
http://<image-service-base-url>/boot-artifacts/rootfs?arch=x86_64&version=4.10

Hence, when deploying the image service, IMAGE_SERVICE_BASE_URL env var
should be specified.

Note: the previous behaviour of using the rootfs_url from OS_IMAGES
is kept as a fallback for now. Can be removed later on once ensuring
IMAGE_SERVICE_BASE_URL is available for the deployment.