NO-ISSUE: Performance enhancements for unit tests
When running unit tests locally, I ran into a couple of issues

1: The tests would have frequent connection failures when trying to create http servers on various ports, this felt like a bug, but would not consistently happen for all users.
2: The unit tests would take a long time to "clean up" adding a lot of uneccesary overhead to tests.

This commit makes sure that when we launch a HTTP or HTTPS server that we wait a reasonable amount of time for them to come online before exiting the non-blocking ListenAndServe() method.
We also added a "Fast shutdown" mode to the HTTP and HTTPS servers where the connection will be closed instead of shut down. This is exclusively for use in testing environments and should not be used in production.