MGMT-11442: Fetch discovery ignition with static networking if minimal iso requested. (#93)
This will make a request to the assisted-service, passing the value of the request parameter "imageType"
When the PR https://github.com/openshift/assisted-service/pull/4309 is merged then the assisted service will respond by
including static networking config in the discovery ignition when a minimal iso is requested.

The static networking config will not be included in the ignition if a full iso is requested.