Include static network config in initrd pxe artifact
Originally it was thought that this would not be necessary as iPXE
requires network access to work. This became an issue when working with
ironic though.

In order to configure iPXE for the hosts it is trying to boot, ironic
controlls DHCP, and therefore, DNS, for those hosts. The problem is that
they don't forward DNS requests to any DNS servers that happen to
already be configured in the environment.

Adding the static networking config to the initrd artifact will enable a
workaround for this issue where the user would provide dns configuration
to the hosts using nmstate even if that configuration was already
present in the environment.

https://issues.redhat.com/browse/MGMT-12189