MGMT-12054: Support discovery kernel arguments for full-iso and minimal-iso When streaming the discovery ISO file, the kernel arguments are included as part of some files: grub.cfg and isolinux.cfg
Each of these files contains an area where the kernel arguments can be
embedded. This are can be identified by the regex "(\n#+) COREOS_KARG_EMBED_AREA".
The arguments should overwrite this are from the beginning and should
not pass the last # (hash sign). To do that, an overlay was created.

In new Openshift versions, there is a file that contains the file names,
while in old versions the file does not exist. If the file exists, the
files to embed the kernel arguments are taken from this config file.
Otherwise, the default files [/EFI/redhat/grub.cfg, /isolinux/isolinux.cfg]
are used.