Validate the volume ID for downloaded ISOs (#105)
Before this change a user could add a non-rhcos/fcos ISO to our image
list and we would fail in unpredictable ways with unhelpful error
messages. This commit explicitly validates the volume ID on the
downloaded ISO to ensure we're dealing with something we support.

With this change the image service fails with a clear error message when
configured with an unsupported ISO:
```
{"file":"/go/src/github.com/openshift/origin/pkg/imagestore/imagestore.go:210","func":"github.com/openshift/assisted-image-service/pkg/imagestore.(*rhcosStore).Populate.func1","level":"error","msg":"failed to validate /data/rhcos-full-iso-4.8-48.84.202109241901-0-x86_64.iso: ISO volume identifier (Fedora-S-dvd-x86_64-37) is invalid","time":"2022-12-02T15:00:52Z"}
{"file":"/go/src/github.com/openshift/origin/main.go:74","func":"main.main.func1","level":"fatal","msg":"Failed to populate image store: failed to validate /data/rhcos-full-iso-4.8-48.84.202109241901-0-x86_64.iso: ISO volume identifier (Fedora-S-dvd-x86_64-37) is invalid\n","time":"2022-12-02T15:00:52Z"}
```

This change also removes the invalid full ISO. Before, the full ISO
would be cached and wouldn't be removed upon failure.

Resolves https://issues.redhat.com/browse/MGMT-12740