Support ppc64le and s390x (#109)
* Skip minimal iso extraction if arch is s390x

There is no way for the image service to edit the kernel parameters for
the s390x ISO which means the rootfs url parameter can't be added. This
makes it impossible to create a minimal ISO template.

https://issues.redhat.com/browse/COS-1924 was opened to address this
issue

* Refactor fixTemplateConfigs to separate grub and isolinux

* Create minimal ISO correctly for ppc64le

Add arch param to CreateMinimalISOTemplate so that we can skip editing
the isolinux.cfg which does not exist for ppc64le.

Also add path to find ppc64le's grub.cfg to the available grub.cfg list

* Add ppc64le and s390x to integration tests

* Fail when trying to embed kargs into s390x image

There is no kernel param embed area in these isos so we shouldn't attempt
to find one

Co-authored-by: Nick Carboni <ncarboni@redhat.com>