Use the ignition info file
If the ignition info file (/coreos/igninfo.json) is present in the ISO
use it to determine where the ignition content should be embedded
instead of assuming it should always go in "/images/ignition.img"

If an error is encountered when reading the ignition info, fall back to
the previous ignition path. This allows the image service to seamlessly
support older versions without having to explicitly include logic about
which version is being used in every call.

Resolves https://issues.redhat.com/browse/MGMT-13692