Add unversioned pre-release/latest isos to integration tests
Previously most of the entries in the integration tests were specific
versions that were not being updated. This commit updates the versions
list to be moving references that will alert us to issues with either
the latest version or pre-release version.

The expectation is that issues with supported versions will be caught by
other e2e tests or in stage when the versions are added.