OCPBUGS-22427: Use 'igninfo.json' to find ignition location
Since version 0.17.0 of the CoreOS installer there is a new
`/coreos/igninfo.json` file that indicates the location of the ignition
configuration inside the ISO. For example, for version 4.14 the S390
live ISO contains this:

```json
{
  "file": "images/cdboot.img",
  "length": 262144,
  "offset": 66497660
}
```

But we currently ignore this file and expect the ignition configuration
always in `/images/ignition.img`, and therefore we fail for these S390
live ISO.

This patch changes the service so that it always looks for the
`igninfo.json` file. If it exists it will use the location it indicates,
if it doesn't exist it will use the default location.

Related: https://issues.redhat.com/browse/OCPBUGS-22427
Related: https://github.com/coreos/coreos-installer/releases/tag/v0.17.0
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>