MGMT-16507: Enable image service to use additional ca certificates (#172)
The infrastructure operator will make available a file at ASSISTED_SERVICE_API_TRUSTED_CA_FILE

This file will contain certificates that are to be used for certificate authority during image fetch
The purpose of this PR is to use those certificates for the fetching of images if they have been set up.

Additionally, we needed to make the  the purpose of HTTPS_CA_FILE clearer, so we added a new variable OS_IMAGE_DOWNLOAD_TRUSTED_CA_FILE
The value of OS_IMAGE_DOWNLOAD_TRUSTED_CA_FILE will default to the value held in HTTPS_CA_FILE unless overridden
The idea is to retire HTTPS_CA_FILE as it is not very clear what it's for.

OSImageDownloadTrustedCAFile is a path to a CA file that will be trusted when fetching OS Images
AssistedServiceApiTrustedCAFile is a path to a CA file that will be trusted for TLS connections to the Assisted Service API