When using a custom CA for OS image download, it was discovered that the default CA's were not being used. (#178)
This is because of how the field `RootCAs` in `TLSClientConfig` is handled in the http client;
if `RootCAs` in `TLSClientConfig` is nil then the default CA's are used, however if the value of `RootCAs` is not nil then the default CA's are overwritten.

The solution is to use x509.SystemCertPool() instead of creating a new one.

This PR ensures that our custom CA is appended to to the default CA's rather than being treated as their replacement.