MGMT-17006: Fix incorrect env var names used for osImageDownloadHeadersMap and osImageDownloadQueryParamsMap (#180)
We recently introduced a feature that allows the passing of headers and query parameters for a request for OS images.
The environment variable names we used in the image service were not consistent with the Volume bindings in
the statefulset generated by the infrastructure operator.

This PR fixes that by applying the correct bindings.

This code as been tested "end to end" in a QE environment.