isoeditor: Allow overflow in IgnitionImageReader()
When getting a reader for the ignition image from IgnitionImageReader(),
don't fail if the ignition size is larger than can fit in the ignition
image that exists in the ISO, provided that no data in the ignition
image outside of the embed area will be overwritten.

This is needed for OCPBUGS-14478.