MULTIARCH-2678, OCPBUGS-14478: Add a NewIgnitionImageReader() function (#185)
* isoeditor: Refactor ignitionBoundariesFinder()

* isoeditor: Refactor NewRHCOSStreamReader()

Create a separate function for performing just the ignition overlay.

* isoeditor: Add NewIgnitionImageReader() function

Return the filename within the ISO and a Reader for the ignition image.
This should work for editing the ignition image even on platforms (like
s390) where the embed area is not the only thing in the ignition image.

This is needed for MULTIARCH-2678.

* isoeditor: Allow overflow in IgnitionImageReader()

When getting a reader for the ignition image from IgnitionImageReader(),
don't fail if the ignition size is larger than can fit in the ignition
image that exists in the ISO, provided that no data in the ignition
image outside of the embed area will be overwritten.

This is needed for OCPBUGS-14478.

* isoeditor: Return updated igninfo.json when expanding ignition image

Ensure that the igninfo.json file matches the actual size of the
ignition area if the ignition area is expanded to accomodate a larger
ignition in NewIgnitionImageReader() by returning new data for it along
with the ignition image.