OCPBUGS-14478: Pad expanded ignition embed area (#190)
* isoeditor: Overwrite unused ignition area with zeroes

This ensures that even if the base ISO already contains a (larger)
ignition, we will get the same output as from a fresh ISO.

* isoeditor: Pad expanded ignition area to ensure free space

If it is possible to expand the ignition embed area, ensure that there
is always 1/4 of the ignition space left free after the ignition. This
ensures the user can always make minor edits to the ignition using
coreos-installer.