MGMT-21276: Add nmstatectl binary for the s390x arch (#475)
The nmstatectl binary is never fetched for the s390x architecture.
This fix addresses the issue by decoupling the extraction of nmstatectl from the minimal ISO flow - since minimal ISO is not supported for s390x. Instead, the binary is now cached, ensuring it is available for s390x as well.

Co-authored-by: Linoy Hadad <lhadad@redhat.com>